refactor: move duplicated `source_to_table` to `utils`
